pyproject.toml
[project]
name = "multi-agent-ai-research"
version = "0.1.0"
description = "Multi-agent study & coding assistant built with LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langchain",
    "langchain-community",
    "langchain-core<0.4.0",
    "langchain-together==0.2.0",
    "langchain-tavily==0.2.0",
    "langgraph==0.2.45",
    "streamlit==1.39.0",
    "python-dotenv==1.0.1",
    "together==1.3.5",
    "tavily-python==0.5.0",
    "httpx==0.27.2",
    "pytest>=7.0.0",
    "langchain-openai",
    "chromadb",
    "pypdf",
    "langchain_text_splitters",
    "sentence-transformers",
    "pydantic>=2.7.0",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]